requirements.txt
blinker==1.7.0
click==8.1.7
coverage==7.5.0
Flask==3.0.3
gunicorn==22.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.1.2
pytest-env==1.1.3
python-dotenv==1.0.1
Werkzeug==3.0.2
